Prototype Game loop and event handling

+++ views/HexTileView.py
@@ -28,17 +28,17 @@ class HexTileView(Sprite):
         x = self.cube.offset_x
         y = self.cube.offset_y
         if y % 2:
-            pos = (self.width*x + self.width*0.5, self.height*0.75*y)
+            pos = (int(self.width*x + self.width*0.5), int(self.height*0.75*y))
         else:
-            pos = (self.width*x, self.height*0.75*y)
+            pos = (int(self.width*x), int(self.height*0.75*y))
         return pos
 
     def calculate_corners(self):
 
-        c1 = (self.center_w, 0)
-        c2 = (self.width, self.height*0.25) 
-        c3 = (self.width, self.height*0.75)
-        c4 = (self.center_w, self.height)
-        c5 = (0, self.height*0.75)
-        c6 = (0, self.height*0.25)
+        c1 = (int(self.center_w), 0)
+        c2 = (int(self.width), int(self.height*0.25))
+        c3 = (int(self.width), int(self.height*0.75))
+        c4 = (int(self.center_w), int(self.height))
+        c5 = (0, int(self.height*0.75))
+        c6 = (0, int(self.height*0.25))
         return (c1, c2, c3, c4, c5, c6)
